OpenDSSC: fix incorrect truncations when using the `Abs` function; affects multiple components
The issue was probably introduced in the original translation from Pascal to C++.

First detected when investigating an issue with Capacitor when providing `R=0` and `XL!=0`, which should initialize R to `0.001*XL` but was resulting in a different value.

git-svn-id: https://svn.code.sf.net/p/electricdss/code@3938 d8739450-1e93-4ef4-a0af-c327d92816ff

+++ General/PriceShape.cpp
@@ -608,7 +608,7 @@ double TPriceShapeObj::GetPrice(double hr)
 					LastValueAccessed = 1;  // Start over from Beginning
 				for(stop = FNumPoints, i = LastValueAccessed + 1; i <= stop; i++)
 				{
-					if(Abs((int) (Hours[i - 1] - hr)) < 0.00001)  // If close to an actual point, just use it.
+					if(Abs((Hours[i - 1] - hr)) < 0.00001)  // If close to an actual point, just use it.
 					{
 						result = PriceValues[i - 1];
 						LastValueAccessed = i;
